Fullstack Nexus Java-Angular - Frontend - modify new employee and edit

+++ 03-phase/user-management-front/src/app/features/employees/employee/employee.ts
@@ -1,80 +1,56 @@
 import { Component, inject, OnInit, signal } from '@angular/core';
 import { CommonModule } from '@angular/common';
-
-import { FormBuilder, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
-
+//import { FormBuilder, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
 import { EmployeeService } from '../services/employee.service';
 import { EmployeeDTO } from '../models/employee.dto';
+import { EmployeeNewEdit } from '../employee-new-edit/employee-new-edit';
 
 
 @Component({
   selector: 'app-employee',
-  imports: [CommonModule, ReactiveFormsModule],
+  imports: [CommonModule, EmployeeNewEdit],
   templateUrl: './employee.html',
   styleUrl: './employee.css',
 })
 export class Employee implements OnInit {
 
   private empService = inject(EmployeeService);
-  private fb = inject(FormBuilder);
+  
   
   employees = signal<EmployeeDTO[]>([]);
-  employeeForm: FormGroup;
-  isEditing = signal<boolean>(false);
-  currentId = signal<number | null>(null);
-
-  constructor() {
-    this.employeeForm = this.fb.group({
-      name: ['', Validators.required],
-      lastName: ['', Validators.required],
-      email: ['', [Validators.required, Validators.email]]
-    });
-  }
+  selectedEmployee = signal<EmployeeDTO | null>(null);
+  showForm = signal<boolean>(false);
 
   ngOnInit() {
      this.load(); 
   }
+  
 
   load() { 
     this.empService.getAll().subscribe(data => this.employees.set(data)); 
   }
 
+
   onDelete(id: number) {
-    this.empService.delete(id).subscribe(() => this.load());
-  }
-
-
-  // Al hacer clic en Editar
-  onEdit(employee: EmployeeDTO) {
-    this.isEditing.set(true);
-    this.currentId.set(employee.id!);
-    this.employeeForm.patchValue(employee); // Carga los datos en los inputs
-  }
-
-  onSave() {
-    if (this.employeeForm.invalid) return;
-
-    const employeeData: EmployeeDTO = this.employeeForm.value;
-
-    if (this.isEditing()) {
-      // UPDATE
-      this.empService.update(this.currentId()!, employeeData).subscribe(() => {
-        this.resetForm();
-        this.load();
-      });
-    } else {
-      // SAVE 
-      this.empService.save(employeeData).subscribe(() => {
-        this.resetForm();
-        this.load();
-      });
+    if(confirm('Delete employee?')) {
+      this.empService.delete(id).subscribe(() => this.load());
     }
   }
 
-  resetForm() {
-    this.isEditing.set(false);
-    this.currentId.set(null);
-    this.employeeForm.reset();
+
+  openNew() {
+    this.selectedEmployee.set(null);
+    this.showForm.set(true);
+  }
+
+  openEdit(employee: EmployeeDTO) {
+    this.selectedEmployee.set(employee);
+    this.showForm.set(true);
+  }
+
+  handleFinished() {
+    this.showForm.set(false);
+    this.load();
   }
 
 }
